Move migration to its own file, continue to support deprecated field.
